requirements.txt
langchain
python-dotenv
langchain-groq
ipykernel
langchain-google-genai 
langchain-community
pypdf
faiss-cpu
structlog
PyMuPDF
pandas
langchain-core[tracing]
pytest
streamlit
sentence_transformers
-e .
